Solved problem 1 in golang - saved by a kiss
My first real program written in Go. I was quite happy that it was
pretty concise and I like Go so far.

As for the maths problem - I started overcomplicating by thinking about
arrays and slices (mainly because I was reading that chapter when I did
this) but remembered one of the best rules in programming - K.I.S.S.